Add app labels to bookmarks and video_config
- PLAT-1642
- PLAT-1643

This works around the deprecation warnings in those tickets. I have not
been able to repro them in devstack, but the other warnings of this
type that I've tracked down have been early-loading to find XBlocks,
which should be fine going forward. If this proves insufficient when
we do the platform 1.11 upgrade we should be able to fix it
by reordering INSTALLED_APPS.